Extract `fromUserInput` helper function
